// lib/api-zod/src/generated/api.ts
export const analyzeSceneBodySceneTextMin = 20;

export const analyzeSceneBodyChallengeModeDefault = false;

export const AnalyzeSceneBody = zod.object({
  "sceneText": zod.string().min(analyzeSceneBodySceneTextMin),
  "resources": zod.string(),
  "budget": zod.string(),
  "equipment": zod.string(),
  "crewExperience": zod.string(),
  "history": zod.array(zod.object({
  "role": zod.enum(['director', 'assistant']),
  "content": zod.string()
})),
  "challengeMode": zod.boolean().default(analyzeSceneBodyChallengeModeDefault),
  "projectId": zod.string(),
  "sceneId": zod.string(),
  "projectType": zod.string().optional()
})